+++ CMakeLists.txt
@@ -26,10 +26,11 @@ DEAL_II_INITIALIZE_CACHED_VARIABLES()
 DEAL_II_QUERY_GIT_INFORMATION(PF_APPLICATIONS)
 
 SET(TARGET_SRC inst.cc)
+FILE(GLOB_RECURSE ADDITIONAL_TARGET_SRC "src/*.cc")
 
 PROJECT(pf-applications)
 
-ADD_LIBRARY(pf-applications  ${TARGET_SRC})
+ADD_LIBRARY(pf-applications ${TARGET_SRC} ${ADDITIONAL_TARGET_SRC})
 
 install(TARGETS pf-applications LIBRARY DESTINATION lib)
